# refactor all framing protocol code into frame