--- a/conn.go
+++ b/conn.go
@@ -18,12 +18,6 @@ import (
 	"time"
 )
 
-const (
-	defaultFrameSize = 4096
-	flagResponse     = 0x80
-	maskVersion      = 0x7F
-)
-
 //JoinHostPort is a utility to return a address string that can be used
 //gocql.Conn to form a connection with a host.
 func JoinHostPort(addr string, port int) string {