Simplify the protocol implementation so that all the logic for handling different protocol versions is now handled inside frame.go. This should make it easier to improve and change the protocols going forward without having to make lots of changes elsewhere.