@Zariel 2015-03-30
@0x6e6562 this is a fairly large change, and took me a while :)

This is a major refactor of how the framing code (frame.go) and connection (conn.go) are interacting, it removes all knowledge of the protocol from the connection and changes how it reads / write frames to the underlying connection.

Im not sure how relevant they are as I didnt spend anytime on optimising anything, but the bench results are interesting.

`benchmark                           old ns/op     new ns/op     delta
BenchmarkWikiCreateSchema-4         184891883     157283266     -14.93%
BenchmarkWikiCreatePages-4          638084        771324        +20.88%
BenchmarkWikiSelectAllPages-4       1624179       2544581       +56.67%
BenchmarkWikiSelectSinglePage-4     315850        408556        +29.35%
BenchmarkWikiSelectPageCount-4      527972        531931        +0.75%

benchmark                           old allocs     new allocs     delta
BenchmarkWikiCreateSchema-4         32             29             -9.38%
BenchmarkWikiCreatePages-4          37             32             -13.51%
BenchmarkWikiSelectAllPages-4       446            1142           +156.05%
BenchmarkWikiSelectSinglePage-4     55             60             +9.09%
BenchmarkWikiSelectPageCount-4      27             25             -7.41%

benchmark                           old bytes     new bytes     delta
BenchmarkWikiCreateSchema-4         11393         1982          -82.60%
BenchmarkWikiCreatePages-4          6389          1699          -73.41%
BenchmarkWikiSelectAllPages-4       50782         50292         -0.96%
BenchmarkWikiSelectSinglePage-4     7509          3243          -56.81%
BenchmarkWikiSelectPageCount-4      5748          1021          -82.24%`