@0x6e6562 2015-04-07
I've had a good look through this change and I've noted the things that may require some follow up - other than that the patch looks pretty good. I can't help thinking that because of the size of this change, there is likely to be subtle bugs we've both missed. I guess end users are likely to tease these out. Because of this, I'm wondering if we temporarily change the top of the README to reflect the fact that we are trying to mature out new framing code and that if people require more tried and tested code, that we provide a tag to the last commit of the previous generation framing code?