@Zariel 2015-04-07
Yeah, I think we should wait for #316 to merge so that we can verify the authentication handshake works correctly.

With regards to a tag, that makes sense.