@Zariel 2015-04-08
Given that #316 is a bit blocked, I ran the tests manually setting the authenticator, found a bug in the type switch for the authentication. Added some more stringing functions to the frames as well.

I think its in a state to merge now @0x6e6562