@Zariel 2015-04-09
@0x6e6562 I made some changes, namely revert back to copying byte slices when parsing frames, which makes the code much easier to reason about. I also changed it to use separate read/write buffers which also makes it easier to reason about.