@0x6e6562 2015-04-09
I'm fine to leap frog #316 to get this going and code wise this LGTM, but I think we should put a highly visible addendum at the top of the README (which we will delete after this code has been tested out enough in the wild), so that people will be able to see that they're dealing with new code. We should also provide a tag to the previous "stable" revision.