@0x6e6562 2015-04-09
>What shall we call the tag?

How about `1st_gen_framing`?