@Zariel 2015-04-09
For the readme, how about this under the website/docs link

Stability
---------
The underlying framing code was rewritten as part of #339 and as such may have
unforseen bugs, if you would like to use the previously known stable framing code the tag `1st_gen_framing` (180456fef0a3c6d02c51dc7211f49b55e9315867) has
been created.