@0x6e6562 2015-04-09
How about:

"The underlying framing code was rewritten as part of #339 and as such may have
unforseen bugs. If you run into a bug related to wire framing, please raise a ticket and we will try to resolve this as soon as we can. If you require a stable version to pin your production app against, we have tagged the previous stable version in source code, so you can build against this. The tag is called 1st_gen_framing (180456f). This note will be removed as the new generation framing code base matures."