@0x6e6562 2015-04-09
Very nice work - this has laid a flexible foundation for the library going forwards :-)